Simplify idle recovery browser coverage
